vttLib: in compile_instructions, warn and skip contour glyphs that contain components in VTT assembly
These are probably composite glyphs that have subsequently been 'decomposed' in
the UFO, but the VTT data still contain the components instructions.
Here we print a warning, drop the assembly data and skip hinting compilation for
those glyphs.
When this happens, either the UFO outlines or the VTT assembly need to be fixed
in the sources, as vttLib cannot resolve the ambiguity.