write maxp data to 'com.github.daltonmaag.vttLib.plist' in data folder
We prefer to keep the maxp hinting data outside the lib.plist, and self contained in the data folder alongside the TSI ttx files.

We still support reading from the old lib keys, but when dumping we will strip those and write to the new vttLib.plist file in the data subfolder.

We don't use TTX format for the maxp since fonttools only support merging full tables, not part of them. But we only want to merge some maxp attributes, not all.
Hence, we dump the values to a custom plist dictionary.